# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Transações Financeiras Simplificada",
        "description": "API para gerenciar contas e transações financeiras",
        "version": "1.0.0"
    },
    "paths": {
        "/api/accounts/balance": {
            "get": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Obtém o saldo da conta do usuário",
                "operationId": "bdcb5773fadd3f1eedf6d5a6c5f73e14",
                "parameters": [
                    {
                        "name": "user-id",
                        "in": "header",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "account_type",
                        "in": "query",
                        "description": "Tipo de conta (PAYMENT)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "PAYMENT"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Saldo obtido com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "balance": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 125.5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro na requisição",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Account type not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/accounts/credit": {
            "post": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Credita um valor na conta destino",
                "operationId": "033c61983256c8a8c70df72adb4984a4",
                "parameters": [
                    {
                        "name": "user-id",
                        "in": "header",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount",
                                    "account_destination",
                                    "transaction_code"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100.5
                                    },
                                    "account_destination": {
                                        "type": "integer",
                                        "example": 123
                                    },
                                    "transaction_code": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Operação realizada com sucesso"
                    },
                    "400": {
                        "description": "Erro na requisição",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Account not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Operação não permitida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Credit not allowed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/transfer": {
            "post": {
                "tags": [
                    "Transfers"
                ],
                "summary": "Transfere um valor entre contas",
                "operationId": "a768977dcdad59041e2c7fdd8259e553",
                "parameters": [
                    {
                        "name": "user-id",
                        "in": "header",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount",
                                    "account_origin",
                                    "account_destination"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 150.75
                                    },
                                    "account_origin": {
                                        "type": "integer",
                                        "example": 123
                                    },
                                    "account_destination": {
                                        "type": "integer",
                                        "example": 456
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Transferência realizada com sucesso"
                    },
                    "400": {
                        "description": "Erro na requisição",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Insufficient balance"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Account": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "PAYMENT"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 123
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Accounts",
            "description": "Accounts"
        },
        {
            "name": "Transfers",
            "description": "Transfers"
        }
    ]
}